+++ CHANGELOG.md
@@ -9,11 +9,13 @@ All notable changes to this project will be documented in this file.
 -   Cat command.
 -   Cowsay command.
 -   TextHelper.ConcatenateSentence method.
+-   yat_context_menu input event keybinding.
 
 ### Changed
 
 -   Godot.NET.Sdk version to 4.2.0-rc.2.
 -   The terminal now distinguishes between sentences wrapped in " or ' and treats them as a whole.
+-   Renamed keybindings for example scenes.
 
 ## [1.9.0-beta 2023-11-23]