+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 -   Quick Commands for terminal.
 -   StartsWith & EndsWith text helper method.
 -   Support for sentences in options.
+-   Terminal documentation.
 
 ### Changed